+++ Angular/src/app/add-role-modal/add-role-modal.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { AddRoleModalComponent } from './add-role-modal.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { DataService } from '../_services/data.service';
+import { DataService } from '../services/data.service';
 
 describe('AddRoleModalComponent', () => {
   let component: AddRoleModalComponent;
