+++ Angular/src/app/app.module.ts
@@ -6,10 +6,9 @@ import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
 
 import { AppComponent } from "./app.component";
 import { LoginComponent } from "./login/login.component";
-import { HttphandlerService } from "./httphandler.service";
-import { AppRoutingModule } from "./app-routing.module";
+import { HttphandlerService } from "./services/httphandler.service";
 import { AuthGuardComponent } from "./auth-guard/auth-guard.component";
-import { DataService } from "./_services/data.service";
+import { DataService } from "./services/data.service";
 import { CareerTableComponent } from "./career-table/career-table.component";
 import { CookieService } from "ngx-cookie-service";
 import { LandingPageComponent } from "./landing-page/landing-page.component";
@@ -20,6 +19,7 @@ import { AddRoleModalComponent } from "./add-role-modal/add-role-modal.component
 import { EditRoleModalComponent } from './edit-role-modal/edit-role-modal.component';
 import { AdminAuthGuardComponent } from './admin-auth-guard/admin-auth-guard.component';
 import { AdminTableComponent } from './admin-table/admin-table.component';
+import { AppRoutingModule } from './app-routing.module';
 import { AddCapabilityComponent } from './add-capability/add-capability.component';
 
 @NgModule({
