+++ Angular/src/app/role-information/role-information.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { RoleInformationComponent } from './role-information.component';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { MockObjects } from '../mock-objects';
+import { MockObjects } from '../services/mock-objects';
 
 describe('RoleInformationComponent', () => {
   let component: RoleInformationComponent;
