+++ Angular/src/app/services/modal.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 
 import { ModalService } from './modal.service';
-import { DataService } from './_services/data.service';
+import { DataService } from './data.service';
 
 describe('ModalService', () => {
   beforeEach(() => TestBed.configureTestingModule({
