+++ Angular/src/app/services/modal.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { DataService } from './_services/data.service';
+import { DataService } from './data.service';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { RoleInformationComponent } from './role-information/role-information.component';
-import { BandInformationComponent } from './band-information/band-information.component';
-import { AddRoleModalComponent } from './add-role-modal/add-role-modal.component';
-import { AddCapabilityComponent } from './add-capability/add-capability.component';
-import { EditRoleModalComponent } from './edit-role-modal/edit-role-modal.component';
+import { RoleInformationComponent } from '../role-information/role-information.component';
+import { BandInformationComponent } from '../band-information/band-information.component';
+import { AddRoleModalComponent } from '../add-role-modal/add-role-modal.component';
+import { AddCapabilityComponent } from '../add-capability/add-capability.component';
+import { EditRoleModalComponent } from '../edit-role-modal/edit-role-modal.component';
 
 @Injectable({
   providedIn: 'root'
